# docs/openapi.yaml
openapi: 3.1.0
info:
  title: darkTunes Music Group API
  description: |
    REST API for the darkTunes Music Group website (Next.js 15 App Router).

    **Authentication**
    - `SessionBearer`: Supabase JWT for authenticated users (portal, account, press).
    - `AdminBearer`: Supabase JWT with `admin` or `editor` role (admin routes).
    - `CronSecret`: Bearer token matching `CRON_SECRET` for scheduled sync/cache jobs.

    **Error format** (all routes via `withErrorHandler`):
    ```json
    { "error": "Human-readable message", "code": "OPTIONAL_ERROR_CODE" }
    ```
  version: 1.0.0
  contact:
    name: darkTunes Music Group
    url: https://darktunes.com
servers:
  - url: https://darktunes.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    SessionBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token (portal artist, press journalist, account owner).
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token with admin or editor role.
    CronSecret:
      type: http
      scheme: bearer
      description: Static secret from CRON_SECRET env var (Supabase Cron / internal jobs).
    PartnerApiKey:
      type: http
      scheme: bearer
      description: >
        Organization API key, prefixed `dt_live_`, sent as `Authorization: Bearer <key>`.
        Keys carry scopes (e.g. `read`, or `*` for all) and are validated against
        organization_api_keys; revoked keys are rejected. The owning organization
        must have the `partner_api` feature enabled.
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Insufficient role or permission
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
paths:
...
  /api/sync-api:
    post:
      operationId: postsync-api
      tags: [Sync]
      summary: Create or submit /api/sync-api
      security:
        - CronSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...